pyproject.toml
[tool.poetry]
name = "khinsider_bot"
version = "0.1.8"
description = "Telegram bot to download soundtracks"
authors = ["NovaHFly <mugiviska@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^22.1"
khinsider-downloader = {git = "https://github.com/NovaHFly/khinsider_downloader"}
starlette = "^0.46.2"
uvicorn = "^0.34.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
